Move port and edge utils out of composables
